Update README.md
Google Chrome blocks the download of the NHML zip file, which may lead some users to think there is a problem with the program (see issue #262 as a perfect example). The proposed addition to the README.md adds an explanation to why this happens and how to get past it.